+++ gpu/thrust.cu
@@ -9,16 +9,16 @@ void sort(const int size, int * key, int * value) {
   thrust::sort_by_key(keyBegin, keyEnd, valueBegin);
 }
 
-void sort(const int size, unsigned long long * key, int * value) {
-  thrust::device_ptr<unsigned long long> keyBegin(key);
-  thrust::device_ptr<unsigned long long> keyEnd(key+size);
+void sort(const int size, uint64_t * key, int * value) {
+  thrust::device_ptr<uint64_t> keyBegin(key);
+  thrust::device_ptr<uint64_t> keyEnd(key+size);
   thrust::device_ptr<int> valueBegin(value);
   thrust::sort_by_key(keyBegin, keyEnd, valueBegin);
 }
 
-void scan(const int size, unsigned long long * key, int * value) {
-  thrust::device_ptr<unsigned long long> keyBegin(key);
-  thrust::device_ptr<unsigned long long> keyEnd(key+size);
+void scan(const int size, uint64_t * key, int * value) {
+  thrust::device_ptr<uint64_t> keyBegin(key);
+  thrust::device_ptr<uint64_t> keyEnd(key+size);
   thrust::device_ptr<int> valueBegin(value);
   thrust::inclusive_scan_by_key(keyBegin, keyEnd, valueBegin, valueBegin);
 }
